Move `FirmwareManagerWorker` in a `worker` module
